Extract server logic into separate file
